requirements.txt
# Core AI and ML Libraries
torch
transformers
peft
bitsandbytes
accelerate

# Data Handling
pandas

# PostgreSQL Database Connector
psycopg2-binary

# Environment Variable Management
python-dotenv
